+++ cglauncher/globals.hpp
@@ -11,6 +11,8 @@ extern HWND hStatic_2_Version;
 extern HWND hStatic_1_Version;
 extern HWND hStatic_x_State;
 extern HWND hStatic_1_Logo;
+extern HWND hNormal_sf_Cancel;
+extern HWND hStatic_sf_UpdateProgress;
 
 extern HFONT mainFont;
 extern HFONT normalFont;